refactor(core): move more host directive matching logic into feature
Moves the logic that tracks host directives into the `ɵɵHostDirectivesFeature` so it can be tree shaken.
