+++ packages/core/src/render3/view/directives.ts
@@ -10,7 +10,6 @@ import {RuntimeError, RuntimeErrorCode} from '../../errors';
 import {Writable} from '../../interface/type';
 import {DoCheck, OnChanges, OnInit} from '../../interface/lifecycle_hooks';
 import {
-  assertDefined,
   assertGreaterThan,
   assertGreaterThanOrEqual,
   assertNotEqual,
@@ -26,6 +25,7 @@ import type {
   DirectiveDef,
   HostDirectiveDef,
   HostDirectiveDefs,
+  HostDirectiveRanges,
 } from '../interfaces/definition';
 import {NodeInjectorFactory} from '../interfaces/injector';
 import {
@@ -52,20 +52,6 @@ export type DirectiveMatcherStrategy = (
   tNode: TElementNode | TContainerNode | TElementContainerNode,
 ) => DirectiveDef<unknown>[] | null;
 
-/** Data produced after host directives are resolved for a node. */
-type HostDirectiveResolution = [
-  matches: DirectiveDef<unknown>[],
-  hostDirectiveDefs: HostDirectiveDefs | null,
-  hostDirectiveRanges: HostDirectiveRanges | null,
-];
-
-/**
- * Map that tracks a selector-matched directive to the range within which its host directives
- * are declared. Host directives for a specific directive are always contiguous within the runtime.
- * Note that both the start and end are inclusive and they're both **after** `tNode.directiveStart`.
- */
-type HostDirectiveRanges = Map<DirectiveDef<unknown>, [start: number, end: number]>;
-
 /**
  * Resolve the matched directives on a node.
  */
@@ -83,15 +69,16 @@ export function resolveDirectives(
   const matchedDirectiveDefs = directiveMatcher(tView, tNode);
 
   if (matchedDirectiveDefs !== null) {
-    let directiveDefs: DirectiveDef<unknown>[];
+    let directiveDefs = matchedDirectiveDefs;
     let hostDirectiveDefs: HostDirectiveDefs | null = null;
     let hostDirectiveRanges: HostDirectiveRanges | null = null;
-    const hostDirectiveResolution = resolveHostDirectives(matchedDirectiveDefs);
 
-    if (hostDirectiveResolution === null) {
-      directiveDefs = matchedDirectiveDefs;
-    } else {
-      [directiveDefs, hostDirectiveDefs, hostDirectiveRanges] = hostDirectiveResolution;
+    for (const def of matchedDirectiveDefs) {
+      if (def.resolveHostDirectives !== null) {
+        [directiveDefs, hostDirectiveDefs, hostDirectiveRanges] =
+          def.resolveHostDirectives(matchedDirectiveDefs);
+        break;
+      }
     }
 
     ngDevMode && assertNoDuplicateDirectives(directiveDefs);
@@ -131,83 +118,6 @@ function cacheMatchingLocalNames(
       );
     localNames.push(localRefs[i], index);
   }
-}
-
-function resolveHostDirectives(matches: DirectiveDef<unknown>[]): HostDirectiveResolution | null {
-  let componentDef: ComponentDef<unknown> | null = null;
-  let hasHostDirectives = false;
-
-  // Having host directives is the less common scenario. Make an initial
-  // validation pass so we don't allocate memory unnecessarily.
-  for (let i = 0; i < matches.length; i++) {
-    const def = matches[i];
-
-    // Given that we may need this further down, we can resolve it already while validating.
-    if (i === 0 && isComponentDef(def)) {
-      componentDef = def;
-    }
-
-    if (def.findHostDirectiveDefs !== null) {
-      hasHostDirectives = true;
-      break;
-    }
-  }
-
-  // If there's at least one def with host directive, we can't bail out of this function.
-  if (!hasHostDirectives) {
-    return null;
-  }
-
-  const allDirectiveDefs: DirectiveDef<unknown>[] = [];
-  let hostDirectiveDefs: HostDirectiveDefs | null = null;
-  let hostDirectiveRanges: HostDirectiveRanges | null = null;
-
-  // Components are inserted at the front of the matches array so that their lifecycle
-  // hooks run before any directive lifecycle hooks. This appears to be for ViewEngine
-  // compatibility. This logic doesn't make sense with host directives, because it
-  // would allow the host directives to undo any overrides the host may have made.
-  // To handle this case, the host directives of components are inserted at the beginning
-  // of the array, followed by the component. As such, the insertion order is as follows:
-  // 1. Host directives belonging to the selector-matched component.
-  // 2. Selector-matched component.
-  // 3. Host directives belonging to selector-matched directives.
-  // 4. Selector-matched dir
-  for (const def of matches) {
-    if (def.findHostDirectiveDefs !== null) {
-      hostDirectiveDefs ??= new Map();
-      hostDirectiveRanges ??= new Map();
-      resolveHostDirectivesForDef(def, allDirectiveDefs, hostDirectiveRanges, hostDirectiveDefs);
-    }
-
-    // Component definition needs to be pushed early to maintain the correct ordering.
-    if (def === componentDef) {
-      allDirectiveDefs.push(def);
-    }
-  }
-
-  if (componentDef === null) {
-    allDirectiveDefs.push(...matches);
-  } else {
-    allDirectiveDefs.push(...matches.slice(1));
-  }
-
-  return [allDirectiveDefs, hostDirectiveDefs, hostDirectiveRanges];
-}
-
-function resolveHostDirectivesForDef(
-  def: DirectiveDef<unknown>,
-  allDirectiveDefs: DirectiveDef<unknown>[],
-  hostDirectiveRanges: HostDirectiveRanges,
-  hostDirectiveDefs: HostDirectiveDefs,
-) {
-  ngDevMode && assertDefined(def.findHostDirectiveDefs, 'Expected host directive resolve function');
-  const start = allDirectiveDefs.length;
-  // TODO(pk): probably could return matches instead of taking in an array to fill in?
-  def.findHostDirectiveDefs!(def, allDirectiveDefs, hostDirectiveDefs);
-
-  // Note that these indexes are within the offset by `directiveStart`. We can't do the
-  // offsetting here, because `directiveStart` hasn't been initialized on the TNode yet.
-  hostDirectiveRanges.set(def, [start, allDirectiveDefs.length - 1]);
 }
 
 /**
