+++ packages/core/test/bundling/animations-standalone/bundle.golden_symbols.json
@@ -443,7 +443,6 @@
   "requiresRefreshOrTraversal",
   "resetPreOrderHookFlags",
   "resolveForwardRef",
-  "resolveHostDirectivesForDef",
   "resolveTiming",
   "resolveTimingValue",
   "roundOffset",
