+++ packages/core/test/bundling/defer/bundle.golden_symbols.json
@@ -845,7 +845,6 @@
   "resetPreOrderHookFlags",
   "resolveDirectives",
   "resolveForwardRef",
-  "resolveHostDirectivesForDef",
   "retrieveHydrationInfo",
   "runEffectsInView",
   "runInInjectionContext",
