+++ packages/core/test/bundling/forms_reactive/bundle.golden_symbols.json
@@ -570,7 +570,6 @@
   "resetPreOrderHookFlags",
   "resolveDirectives",
   "resolveForwardRef",
-  "resolveHostDirectivesForDef",
   "resolveProvider",
   "runEffectsInView",
   "runInInjectionContext",
