+++ packages/core/test/bundling/forms_template_driven/bundle.golden_symbols.json
@@ -561,7 +561,6 @@
   "resetPreOrderHookFlags",
   "resolveDirectives",
   "resolveForwardRef",
-  "resolveHostDirectivesForDef",
   "resolveProvider",
   "resolvedPromise",
   "resolvedPromise2",
