+++ packages/core/test/bundling/hydration/bundle.golden_symbols.json
@@ -408,7 +408,6 @@
   "requiresRefreshOrTraversal",
   "resetPreOrderHookFlags",
   "resolveForwardRef",
-  "resolveHostDirectivesForDef",
   "retrieveHydrationInfo",
   "retrieveHydrationInfoImpl",
   "runEffectsInView",
