+++ packages/core/test/bundling/router/bundle.golden_symbols.json
@@ -651,7 +651,6 @@
   "requiresRefreshOrTraversal",
   "resetPreOrderHookFlags",
   "resolveForwardRef",
-  "resolveHostDirectivesForDef",
   "rootRoute",
   "routes",
   "runEffectsInView",
