+++ packages/core/test/bundling/standalone_bootstrap/bundle.golden_symbols.json
@@ -340,7 +340,6 @@
   "requiresRefreshOrTraversal",
   "resetPreOrderHookFlags",
   "resolveForwardRef",
-  "resolveHostDirectivesForDef",
   "runEffectsInView",
   "runInInjectionContext",
   "saveNameToExportMap",
