+++ packages/core/test/bundling/todo/bundle.golden_symbols.json
@@ -454,7 +454,6 @@
   "resetPreOrderHookFlags",
   "resolveDirectives",
   "resolveForwardRef",
-  "resolveHostDirectivesForDef",
   "runEffectsInView",
   "runInInjectionContext",
   "saveNameToExportMap",
